:bug: make id from short to int

+++ src/model/Task.java
@@ -12,18 +12,18 @@ public class Task {
     private boolean requirePhysicalAppearance;
     private Location location;
     private int id;
-    private short idFirstVirtual;
-    private short idSecondVisit;
-    private short idSecondVisitVirtual;
+    private int idFirstVirtual;
+    private int idSecondVisit;
+    private int idSecondVisitVirtual;
     private boolean prioritized;
     private int transportType;     // 0 - drive, 1 - walk
     private int transportedBy;
 
-    public Task(int id, short numTasks ){
+    public Task(int id, int numTasks ){
         this.id = id;
-        this.idFirstVirtual = (short) (id + 2 * numTasks);
-        this.idSecondVisit = (short) (id + numTasks);
-        this.idSecondVisitVirtual = (short) (id + 3 * numTasks);
+        this.idFirstVirtual = ( int) (id + 2 * numTasks);
+        this.idSecondVisit = ( int) (id + numTasks);
+        this.idSecondVisitVirtual = ( int) (id + 3 * numTasks);
 
     };
     
@@ -69,19 +69,19 @@ public class Task {
 
     public int getTransportedBy(){ return this.transportedBy; }
 
-    public short getFirstVisitVirtualId() {
+    public  int getFirstVisitVirtualId() {
         return this.idFirstVirtual;
     }
 
-    public short getSecondVisitVirtualId() {
+    public  int getSecondVisitVirtualId() {
         return this.idSecondVisitVirtual;
     }
 
-    public short getFirstVisitId() {
+    public int getFirstVisitId() {
         return this.id;
     }
 
-    public short getSecondVisitId() {
+    public  int getSecondVisitId() {
         return this.idSecondVisit;
     }
 }
